move Rank, Suit, Card, and Deck into corresponding Base classes
