+++ lib/cardshark/rank.rb
@@ -1,10 +1,6 @@
 # frozen_string_literal: true
 
-require 'cardshark/error'
-require 'cardshark/dimension'
-
 module Cardshark
-  class Rank
-    include Dimension
+  module Rank
   end
 end
