+++ lib/cardshark/suit.rb
@@ -1,10 +1,6 @@
 # frozen_string_literal: true
 
-require 'cardshark/error'
-require 'cardshark/dimension'
-
 module Cardshark
-  class Suit
-    include Dimension
+  module Suit
   end
 end
