+++ lib/cardshark/suit/base.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require 'cardshark/error'
+require 'cardshark/dimension'
+
+module Cardshark
+  module Suit
+    class Base
+      include Dimension
+    end
+  end
+end
