+++ spec/cardshark/rank/base_spec.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'cardshark/rank/base'
+
+RSpec.describe Cardshark::Rank::Base do
+end
