--- spec/cardshark/rank_spec.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'cardshark/rank'
-
-RSpec.describe Cardshark::Rank do
-end
