+++ spec/cardshark/suit/base_spec.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'cardshark/suit/base'
+
+RSpec.describe Cardshark::Suit::Base do
+end
